first implementation, has race conditions with resource transactionflow
