+++ auxo/resource_protectsurface.go
@@ -2,15 +2,12 @@ package auxo
 
 import (
 	"context"
-	"strings"
-	"time"
 
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/booldefault"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/int64default"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/int64planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/mapplanmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/setplanmodifier"
@@ -30,43 +27,29 @@ type protectsurfaceResource struct {
 }
 
 type protectsurfaceResourceModel struct {
-	ID                    types.String       `tfsdk:"id"`
-	Uniqueness_key        types.String       `tfsdk:"uniqueness_key"`
-	Name                  types.String       `tfsdk:"name"`
-	Description           types.String       `tfsdk:"description"`
-	MainContact           types.String       `tfsdk:"main_contact"`
-	SecurityContact       types.String       `tfsdk:"security_contact"`
-	InControlBoundary     types.Bool         `tfsdk:"in_control_boundary"`
-	InZeroTrustFocus      types.Bool         `tfsdk:"in_zero_trust_focus"`
-	Relevance             types.Int64        `tfsdk:"relevance"`
-	Confidentiality       types.Int64        `tfsdk:"confidentiality"`
-	Integrity             types.Int64        `tfsdk:"integrity"`
-	Availability          types.Int64        `tfsdk:"availability"`
-	DataTags              types.Set          `tfsdk:"data_tags"`
-	ComplianceTags        types.Set          `tfsdk:"compliance_tags"`
-	CustomerLabels        types.Map          `tfsdk:"customer_labels"`
-	SOCTags               types.Set          `tfsdk:"soc_tags"`
-	AllowFlowsFromOutside types.Bool         `tfsdk:"allow_flows_from_outside"`
-	AllowFlowsToOutside   types.Bool         `tfsdk:"allow_flows_to_outside"`
-	MaturityStep1         types.Int64        `tfsdk:"maturity_step1"`
-	MaturityStep2         types.Int64        `tfsdk:"maturity_step2"`
-	MaturityStep3         types.Int64        `tfsdk:"maturity_step3"`
-	MaturityStep4         types.Int64        `tfsdk:"maturity_step4"`
-	MaturityStep5         types.Int64        `tfsdk:"maturity_step5"`
-	Measures              map[string]measure `tfsdk:"measures"`
-}
-
-type measure struct {
-	//Measure               types.String `tfsdk:"measure"`
-	Assigned              types.Bool   `tfsdk:"assigned"`
-	Assigned_by           types.String `tfsdk:"assigned_by"`
-	Assigned_timestamp    types.Int64  `tfsdk:"assigned_timestamp"`
-	Implemented           types.Bool   `tfsdk:"implemented"`
-	Implemented_by        types.String `tfsdk:"implemented_by"`
-	Implemented_timestamp types.Int64  `tfsdk:"implemented_timestamp"`
-	Evidenced             types.Bool   `tfsdk:"evidenced"`
-	Evidenced_by          types.String `tfsdk:"evidenced_by"`
-	Evidenced_timestamp   types.Int64  `tfsdk:"evidenced_timestamp"`
+	ID                    types.String `tfsdk:"id"`
+	Uniqueness_key        types.String `tfsdk:"uniqueness_key"`
+	Name                  types.String `tfsdk:"name"`
+	Description           types.String `tfsdk:"description"`
+	MainContact           types.String `tfsdk:"main_contact"`
+	SecurityContact       types.String `tfsdk:"security_contact"`
+	InControlBoundary     types.Bool   `tfsdk:"in_control_boundary"`
+	InZeroTrustFocus      types.Bool   `tfsdk:"in_zero_trust_focus"`
+	Relevance             types.Int64  `tfsdk:"relevance"`
+	Confidentiality       types.Int64  `tfsdk:"confidentiality"`
+	Integrity             types.Int64  `tfsdk:"integrity"`
+	Availability          types.Int64  `tfsdk:"availability"`
+	DataTags              types.Set    `tfsdk:"data_tags"`
+	ComplianceTags        types.Set    `tfsdk:"compliance_tags"`
+	CustomerLabels        types.Map    `tfsdk:"customer_labels"`
+	SOCTags               types.Set    `tfsdk:"soc_tags"`
+	AllowFlowsFromOutside types.Bool   `tfsdk:"allow_flows_from_outside"`
+	AllowFlowsToOutside   types.Bool   `tfsdk:"allow_flows_to_outside"`
+	MaturityStep1         types.Int64  `tfsdk:"maturity_step1"`
+	MaturityStep2         types.Int64  `tfsdk:"maturity_step2"`
+	MaturityStep3         types.Int64  `tfsdk:"maturity_step3"`
+	MaturityStep4         types.Int64  `tfsdk:"maturity_step4"`
+	MaturityStep5         types.Int64  `tfsdk:"maturity_step5"`
 }
 
 func NewProtectsurfaceResource() resource.Resource {
@@ -266,82 +249,6 @@ func (r *protectsurfaceResource) Schema(ctx context.Context, req resource.Schema
 				Computed:            true,
 				Default:             int64default.StaticInt64(1),
 			},
-			"measures": schema.MapNestedAttribute{
-				Description:         "Measures of the resource protectsurface",
-				MarkdownDescription: "Measures of the resource protectsurface",
-				Optional:            true,
-				NestedObject: schema.NestedAttributeObject{
-					Attributes: map[string]schema.Attribute{
-						"assigned": schema.BoolAttribute{
-							Description:         "Measure assigned to the protectsurface",
-							MarkdownDescription: "Measure assigned to the protectsurface",
-							Required:            true,
-						},
-						"assigned_by": schema.StringAttribute{
-							Description:         "Who assigned this measure to the protectsurface",
-							MarkdownDescription: "Who assigned this measure to the protectsurface",
-							Optional:            true,
-							Computed:            true,
-							Default:             stringdefault.StaticString(""),
-						},
-						"assigned_timestamp": schema.Int64Attribute{
-							Description:         "When was this measure assigned to the protectsurface",
-							MarkdownDescription: "When was this measure assigned to the protectsurface",
-							Optional:            true,
-							Computed:            true,
-							PlanModifiers: []planmodifier.Int64{
-								int64planmodifier.UseStateForUnknown(),
-							},
-						},
-						"implemented": schema.BoolAttribute{
-							Description:         "Is this measure implemented to the protectsurface",
-							MarkdownDescription: "Is this measure implemented to the protectsurface",
-							Optional:            true,
-							Computed:            true,
-							Default:             booldefault.StaticBool(false),
-						},
-						"implemented_by": schema.StringAttribute{
-							Description:         "Who implemented this measure to the protectsurface",
-							MarkdownDescription: "Who implemented this measure to the protectsurface",
-							Optional:            true,
-							Computed:            true,
-							Default:             stringdefault.StaticString(""),
-						},
-						"implemented_timestamp": schema.Int64Attribute{
-							Description:         "When was this measure implemented to the protectsurface",
-							MarkdownDescription: "When was this measure implemented to the protectsurface",
-							Optional:            true,
-							Computed:            true,
-							PlanModifiers: []planmodifier.Int64{
-								int64planmodifier.UseStateForUnknown(),
-							},
-						},
-						"evidenced": schema.BoolAttribute{
-							Description:         "Is there evidence that this measure is implemented",
-							MarkdownDescription: "Is there evidence that this measure is implemented",
-							Optional:            true,
-							Computed:            true,
-							Default:             booldefault.StaticBool(false),
-						},
-						"evidenced_by": schema.StringAttribute{
-							Description:         "Who evidenced that this measure is implementd",
-							MarkdownDescription: "Who evidenced that this measure is implementd",
-							Optional:            true,
-							Computed:            true,
-							Default:             stringdefault.StaticString(""),
-						},
-						"evidenced_timestamp": schema.Int64Attribute{
-							Description:         "When was this measure evidenced",
-							MarkdownDescription: "When was this measure evidenced",
-							Optional:            true,
-							Computed:            true,
-							PlanModifiers: []planmodifier.Int64{
-								int64planmodifier.UseStateForUnknown(),
-							},
-						},
-					},
-				},
-			},
 		},
 	}
 }
@@ -478,43 +385,6 @@ func (r *protectsurfaceResource) Delete(ctx context.Context, req resource.Delete
 
 }
 
-func (r *protectsurfaceResource) getAvailableMeasures() []string {
-	availableMeasures, _ := r.client.ZeroTrust.GetMeasures()
-	availableMeasuresInSlice := make([]string, 0)
-	for _, mg := range availableMeasures.Groups {
-		for _, m := range mg.Measures {
-			availableMeasuresInSlice = append(availableMeasuresInSlice, m.Name)
-		}
-	}
-
-	return availableMeasuresInSlice
-}
-
-func getMeasuresFromMap(measureMap map[string]zerotrust.MeasureState) map[string]measure {
-	if len(measureMap) == 0 {
-		return nil
-	}
-
-	measures := make(map[string]measure, len(measureMap))
-
-	for k, state := range measureMap {
-		measures[k] = measure{
-			//Measure:               types.StringValue(m),
-			Assigned:              types.BoolValue(state.Assignment.Assigned),
-			Assigned_by:           types.StringValue(state.Assignment.LastDeterminedByPersonID),
-			Assigned_timestamp:    types.Int64Value(int64(state.Assignment.LastDeterminedTimestamp)),
-			Implemented:           types.BoolValue(state.Implementation.Implemented),
-			Implemented_by:        types.StringValue(state.Implementation.LastDeterminedByPersonID),
-			Implemented_timestamp: types.Int64Value(int64(state.Implementation.LastDeterminedTimestamp)),
-			Evidenced:             types.BoolValue(state.Evidence.Evidenced),
-			Evidenced_by:          types.StringValue(state.Evidence.LastDeterminedByPersonID),
-			Evidenced_timestamp:   types.Int64Value(int64(state.Evidence.LastDeterminedTimestamp)),
-		}
-	}
-
-	return measures
-}
-
 // resourceModelToProtectsurface maps the resource model to the zerotrust.protectsurface object
 func resourceModelToProtectsurface(plan *protectsurfaceResourceModel, ctx context.Context, r *protectsurfaceResource) (zerotrust.ProtectSurface, diag.Diagnostics) {
 	var diag diag.Diagnostics
@@ -565,69 +435,6 @@ func resourceModelToProtectsurface(plan *protectsurfaceResourceModel, ctx contex
 		},
 	}
 
-	measureMap := make(map[string]zerotrust.MeasureState, 0)
-
-	//Loop through measures
-	for k, m := range plan.Measures {
-
-		//Check if measure exists
-		if !sliceContains(r.getAvailableMeasures(), k) {
-			diag.AddError("Measure does not exists.",
-				"Messure ["+k+"] does not exist, available measures ["+strings.Join(r.getAvailableMeasures(), ",")+"]")
-			return zerotrust.ProtectSurface{}, diag
-		}
-
-		var assigned_timestamp int
-		if !(m.Assigned_timestamp.IsUnknown() || m.Assigned_timestamp.IsNull()) {
-			assigned_timestamp = int(m.Assigned_timestamp.ValueInt64())
-		} else {
-			assigned_timestamp = int(time.Now().Unix())
-		}
-
-		assignment := zerotrust.Assignment{
-			Assigned:                 m.Assigned.ValueBool(),
-			LastDeterminedByPersonID: m.Assigned_by.ValueString(),
-			LastDeterminedTimestamp:  assigned_timestamp,
-		}
-
-		var implemented_timestamp int
-		if !(m.Implemented_timestamp.IsUnknown() || m.Implemented_timestamp.IsNull()) {
-			implemented_timestamp = int(m.Implemented_timestamp.ValueInt64())
-		} else {
-			implemented_timestamp = int(time.Now().Unix())
-		}
-
-		implementation := zerotrust.Implementation{
-			Implemented:              m.Implemented.ValueBool(),
-			LastDeterminedByPersonID: m.Implemented_by.ValueString(),
-			LastDeterminedTimestamp:  implemented_timestamp,
-		}
-
-		var evidenced_timestamp int
-		if !(m.Evidenced_timestamp.IsUnknown() || m.Evidenced_timestamp.IsNull()) {
-			evidenced_timestamp = int(m.Evidenced_timestamp.ValueInt64())
-		} else {
-			evidenced_timestamp = int(time.Now().Unix())
-		}
-
-		evidence := zerotrust.Evidence{
-			Evidenced:                m.Evidenced.ValueBool(),
-			LastDeterminedByPersonID: m.Evidenced_by.ValueString(),
-			LastDeterminedTimestamp:  evidenced_timestamp,
-		}
-
-		measureMap[k] = zerotrust.MeasureState{
-			Assignment:     &assignment,
-			Implementation: &implementation,
-			Evidence:       &evidence,
-		}
-	}
-
-	if len(measureMap) == 0 {
-		measureMap = nil
-	}
-	protectsurface.Measures = measureMap
-
 	return protectsurface, diag
 }
 
@@ -670,7 +477,6 @@ func protectsurfaceToResourceModel(ps *zerotrust.ProtectSurface, ctx context.Con
 		MaturityStep3:         types.Int64Value(int64(ps.Maturity.Step3)),
 		MaturityStep4:         types.Int64Value(int64(ps.Maturity.Step4)),
 		MaturityStep5:         types.Int64Value(int64(ps.Maturity.Step5)),
-		Measures:              getMeasuresFromMap(ps.Measures),
 	}
 
 	return psrm, diag
