+++ templates/resources/transactionflow.md.tmpl
@@ -17,3 +17,4 @@ description: |-
 
 {{ tffile "examples/resources/transactionflow-bidirectional.tf" }}
 
+{{ .SchemaMarkdown | trimspace }}
